- Separated Community, JoinRequest, and CommunityMember in dedicated folders for clarity - Use of permissions to only allow admin and authenticated users to manipulate JoinRequests through views - A community admin can now list, accept and decline the join requests of the community
