+++ backend/community/api/urls.py
@@ -1,17 +1,18 @@
 from django.urls import path, include
 from rest_framework.routers import DefaultRouter
-from community.api import views
-from rest_framework_nested import routers
+from community.api.joinrequest.views import JoinRequestList, JoinRequestAccept, JoinRequestDecline
+from community.api.community.views import CommunityViewSet
+
 
 app_name = 'community'
 
 router = DefaultRouter()
-router.register(r'', views.CommunityViewSet)
+router.register(r'', CommunityViewSet)
 
 # The API URLs are now determined automatically by the router.
 urlpatterns = [
     path('', include((router.urls))),
-    path('<slug:community_slug>/join-requests', views.JoinRequestList.as_view(), name='join-request-list'),
-    path('<slug:community_slug>/join-requests/<int:join_request_pk>/accept', views.JoinRequestAccept.as_view(), name='join-request-accept'),
-    path('<slug:community_slug>/join-requests/<int:join_request_pk>/decline', views.JoinRequestDecline.as_view(), name='join-request-decline'),
+    path('<slug:community_slug>/join-requests', JoinRequestList.as_view(), name='join-request-list'),
+    path('<slug:community_slug>/join-requests/<int:join_request_pk>/accept', JoinRequestAccept.as_view(), name='join-request-accept'),
+    path('<slug:community_slug>/join-requests/<int:join_request_pk>/decline', JoinRequestDecline.as_view(), name='join-request-decline'),
 ]
